move customized_accuracy
