+++ src/ore_algebra/analytic/factorization.py
@@ -44,12 +44,12 @@ from .monodromy import _monodromy_matrices
 from .differential_operator import PlainDifferentialOperator
 from .accuracy import PrecisionError
 from .complex_optimistic_field import ComplexOptimisticField
-from .utilities import (customized_accuracy, power_series_coerce, derivatives,
+from .utilities import (power_series_coerce, derivatives,
                         hp_approximants, guess_exact_numbers,
                         guess_rational_numbers, guess_algebraic_numbers,
                         euler_representation)
 from .linear_algebra import (invariant_subspace, row_echelon_form, ker,
-                             gen_eigenspaces, orbit)
+                             gen_eigenspaces, orbit, customized_accuracy)
 
 
 Radii = RealField(30)
